Create eip-governance-transparency.md
The motivation for this EIP is faster orientation for the public to the governance process. A new observer has a difficult time understanding how certain people or teams are related to the process. It's difficult to find a place to contribute effectively if roles aren't clearly delineated and existing processes defined.

I think what's written up here is well established within a few different repos but is hard for new people to get straight when trying to understand how Ethereum governance works. Only a couple 'new' things are proposed here - just transparency into *who* relevant parties are.

This EIP serves to keep these practices consistent and to be a reference for those seeking to understand the process.